Forgot password, recaptha vs. gibi user/role ve user-password islemlerini baska branch uzerinde gelistirecegim

+++ src/main/java/com/scalx/devnews/entity/Recommendation.java
@@ -11,7 +11,7 @@ import javax.persistence.Table;
 @Data
 public class Recommendation extends AbstractAuditable {
 
-    private int articleId;
+    private Long articleId;
     private double similarityScore;
 
 //    private boolean isNew;
