+++ src/main/java/com/scalx/devnews/repository/LikeRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface LikeRepository extends JpaRepository<Like, Integer> {
+public interface LikeRepository extends JpaRepository<Like, Long> {
 
     // method needs any param ? bool true/false
     public List<Like> findByIsNew();
