+++ src/main/java/com/scalx/devnews/repository/UrlRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface UrlRepository extends JpaRepository<Url, Integer> {
+public interface UrlRepository extends JpaRepository<Url, Long> {
 
     public List<Url> findByIsNew();
 }
